move message sending into notifications/sender.js
